requirements.txt
streamlit
pandas
duckdb
langchain
langchain-community
openai
python-dotenv
mysql-connector-python
chardet